Backport 4b9ec8245187a2eaccc711a6e5d3d4915dd022c9
